fix/rnn: remove cuda parts from native add TODOs for unimplemented
