+++ coaster-nn/src/plugin.rs
@@ -1,6 +1,5 @@
 //! Provides the INn Plugin trait for Coaster implementation.
 use crate::co::tensor::SharedTensor;
-use crate::cudnn::{FilterDescriptor, TensorDescriptor};
 
 #[derive(Debug, Copy, Clone)]
 /// Different algorithms to compute the convolution forward algorithm.
@@ -151,9 +150,9 @@ pub trait ConvolutionConfig<F> {
 ///
 /// Needs to be implemented for Operation specific configurations.
 pub trait RnnConfig<F> {
-    /// Returns the largest workspace size in bytes needed
-    /// for any of the convolution operations.
-    fn workspace_size(&self) -> usize {0}
+    /// Workspace Size - Overwritten by each plugin method except native, which doesn't require
+    /// a workspace size.
+    fn workspace_size(&self) -> usize { 0 }
 }
 
 /// Provides the functionality for a backend to support Neural Network related operations.
@@ -276,38 +275,6 @@ pub trait TanhPointwise<F> : NN<F> {
     /// Saves the result back to `x_diff`.
     fn tanh_pointwise_grad(&self, x: &SharedTensor<F>, x_diff: &mut SharedTensor<F>)
                            -> Result<(), crate::co::error::Error>;
-}
-
-#[derive(Debug)]
-// All RNN Sequence Descriptors are generated on a single pass in CUDNN example code
-// As such, defining them all in one function appears to be the simplest method of reproducing
-// this work in Rust, but passing back a tuple is unwieldy as the tuple grows beyond 2 - 3 values.
-/// Struct to hold all Sequence Descriptors for an RNN Pass
-pub struct RnnSequenceDescriptors {
-    /// Input Descriptor
-    pub x_desc: Vec<TensorDescriptor>,
-    /// Output Descriptor
-    pub y_desc: Vec<TensorDescriptor>,
-    /// Gradient Input Descriptor
-    pub dx_desc: Vec<TensorDescriptor>,
-    /// Gradient Output Descriptor
-    pub dy_desc: Vec<TensorDescriptor>,
-    /// Hidden Input Descriptor
-    pub hx_desc: TensorDescriptor,
-    /// Cell Input Descriptor
-    pub cx_desc: TensorDescriptor,
-    /// Hidden Output Descriptor
-    pub hy_desc: TensorDescriptor,
-    /// Cell Output Descriptor
-    pub cy_desc: TensorDescriptor,
-    /// Gradient Hidden Input Descriptor
-    pub dhx_desc: TensorDescriptor,
-    /// Gradient Cell Input Descriptor
-    pub dcx_desc: TensorDescriptor,
-    /// Gradient Hidden Output Descriptor
-    pub dhy_desc: TensorDescriptor,
-    /// Gradient Cell Output Descriptor
-    pub dcy_desc: TensorDescriptor,
 }
 
 /// Provide the functionality for a Backend to support RNN operations
